+++ src/omi/license.py
@@ -4,6 +4,8 @@ import json
 import re
 from pathlib import Path
 
+from omi.base import get_metadata_version
+
 LICENCES_FILE = Path(__file__).parent / "data" / "licenses.json"
 
 
@@ -82,27 +84,45 @@ def validate_oemetadata_licenses(metadata: dict) -> None:
     Returns
     -------
     None
-        if licences are valid, otherwise LicenseError is raised
+        if licenses are valid, otherwise LicenseError is raised
     """
     if metadata is None:
         msg = "Metadata is empty."
         raise LicenseError(msg)
 
-    licenses = metadata.get("licenses", [])
+    version = get_metadata_version(metadata)
+    licenses_info = _find_license_field(metadata, version)
 
-    if not licenses:
-        msg = "No license information available in the metadata."
-        raise LicenseError(msg)
+    for resource_index, licenses in licenses_info:
+        if not licenses:
+            raise LicenseError(f"No license information available in the metadata for resource: {resource_index + 1}.")
+        for i, license_ in enumerate(licenses or []):
+            if not license_.get("name"):
+                raise LicenseError(
+                    f"The license name is missing in resource {resource_index + 1}, license {i + 1} ({license_}).",
+                )
 
-    for i, license_ in enumerate(licenses):
-        if not license_.get("name"):
-            raise LicenseError(f"The license name is missing in {i}. license ({license_})")
+            if not validate_license(license_["name"]):
+                raise LicenseError(
+                    f"The (normalized) license name '{license_['name']}' in resource"
+                    f"{resource_index + 1}, license {i + 1} "
+                    "was not found in the SPDX licenses list. "
+                    "(See https://github.com/spdx/license-list-data/blob/main/json/licenses.json).",
+                )
 
-        if not validate_license(license_["name"]):
-            raise LicenseError(
-                f"The (normalized) license name '{license_['name']}' was not found in the SPDX licenses list. "
-                "(See https://github.com/spdx/license-list-data/blob/main/json/licenses.json).",
-            )
+
+def _find_license_field(metadata: dict, version: str) -> list:
+    version = get_metadata_version(metadata)
+    if version == "OEMetadata-2.0.1":
+        # Include resource index with each license for traceability
+        licenses_per_resource = [
+            (i, resource.get("licenses")) for i, resource in enumerate(metadata.get("resources", []))
+        ]
+    else:
+        # Return -1 as a placeholder index for top-level licenses
+        licenses_per_resource = [(0, metadata.get("licenses", []))]
+
+    return licenses_per_resource
 
 
 LICENSES = read_licenses()